+++ database/src/main.py
@@ -1,3 +1,8 @@
+import sys
+import os
+
+sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+
 from fastapi import FastAPI, Depends
 from sqlalchemy.orm import Session
 from src.models import Base, Deployment
